+++ recipe/config/current.php
@@ -13,13 +13,20 @@ use Symfony\Component\Console\Helper\Table;
 desc('Show current paths');
 task('config:current', function () {
     $rows = [];
-    $hosts = Deployer::get()->hostSelector->getHosts(input()->getArgument('stage'));
+    $hosts = Deployer::get()->hosts;
 
     on($hosts, function (Host $host) use (&$rows) {
-        $rows[] = [
-            $host->getHostname(),
-            basename($host->getConfig()->get('current_path')),
-        ];
+        try {
+            $rows[] = [
+                $host->getHostname(),
+                basename($host->getConfig()->get('current_path')),
+            ];
+        } catch (\Throwable $e) {
+            $rows[] = [
+                $host->getHostname(),
+                'unknown',
+            ];
+        }
     });
 
     $table = new Table(output());