+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 # Changelog
 
+
 ## master
-[v5.1.1...master](https://github.com/deployphp/deployer/compare/v5.1.1...master)
+[v5.1.2...master](https://github.com/deployphp/deployer/compare/v5.1.2...master)
+
+
+## v5.1.2
+[v5.1.1...v5.1.2](https://github.com/deployphp/deployer/compare/v5.1.1...v5.1.2)
 
 ### Changed
 - Improved `config:current` output (print each host's current release)